Code refactoring: - Genome moved to separate file (genome.py) - lose functions moved to utils.py - sorted functions in cli.py - fixed most warnings - cleaned requirements - cleaned .travis.yaml
