+++ .travis.yml
@@ -10,8 +10,6 @@ env:
     - CC_TEST_REPORTER_ID=951f438ac8a0fa93801ff0bf69922df59fe03800bf7ea8ab77a3c26cda444979
   jobs:
     - PYTHON_VERSION: "3.6"
-#    - PYTHON_VERSION: "3.7"
-#    - PYTHON_VERSION: "3.8"
 
 before_install:
   # install miniconda
@@ -26,7 +24,6 @@ before_install:
   - conda config --set always_yes yes
 
 install:
-  # python version is specified in environment.yml
   - conda env create -n genomepy python=$PYTHON_VERSION -f environment.yml
   - source activate genomepy
   - python setup.py develop
@@ -36,17 +33,19 @@ before_script:
   # install codeclimate test coverage
   - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then
         wget -O cc-test-reporter https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64;
-    else
-        wget -O cc-test-reporter https://codeclimate.com/downloads/test-reporter/test-reporter-latest-darwin-amd64;
+        chmod +x ./cc-test-reporter;
+        ./cc-test-reporter before-build;
+#    else
+#        wget -O cc-test-reporter https://codeclimate.com/downloads/test-reporter/test-reporter-latest-darwin-amd64;
     fi
-  - chmod +x ./cc-test-reporter
-  - ./cc-test-reporter before-build
+#  - chmod +x ./cc-test-reporter
+#  - ./cc-test-reporter before-build
 
 script:
   - black --check setup.py genomepy/ tests/
   - flake8 setup.py genomepy/ tests/
   - pytest -v --disable-pytest-warnings --cov=genomepy --cov-report=xml tests/
 
-# Pipe the coverage data to Code Climate
 after_script:
+  # send the coverage data to Code Climate
   - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then ./cc-test-reporter after-build -t coverage.py --exit-code $TRAVIS_TEST_RESULT; fi
