+++ genomepy/__init__.py
@@ -1,10 +1,13 @@
 """Search, download and use genome FASTA files."""
 from genomepy.functions import (  # noqa: F401
+    install_genome,
     list_available_providers,
     list_available_genomes,
     list_installed_genomes,
+    manage_config,
+    manage_plugins,
     search,
-    install_genome,
-    Genome,
 )
+from genomepy.genome import Genome  # noqa: F401
+from genomepy.provider import ProviderBase  # noqa: F401
 from genomepy.__about__ import __version__, __author__  # noqa: F401
